Move scaling strategy into jobs module
